src/core/memory_manager.py
"""
Memory manager module for handling memory operations.
"""

import os
import struct
import logging
from typing import Dict, List, Optional, Union
import ctypes
from ctypes import wintypes, c_void_p, c_size_t, c_ulong, c_ulonglong
from keystone import *
from capstone import *

class MemoryManager:
    """Manager for memory operations and patching."""
    
    def __init__(self):
        self.logger = logging.getLogger(__name__)
        self.patches = {}
        self.original_bytes = {}
        self.assembler = None
        self.disassembler = None
        
    def initialize(self, architecture: str = 'x86', mode: int = 32) -> bool:
        """Initialize the memory manager with specified architecture.
        
        Args:
            architecture: Target architecture ('x86', 'arm', 'arm64')
            mode: Architecture mode (32 or 64)
            
        Returns:
            bool: True if initialization was successful
        """
        try:
            # Initialize assembler and disassembler based on architecture
            if architecture == 'x86':
                if mode == 32:
                    self.assembler = self._init_x86_assembler(32)
                    self.disassembler = self._init_x86_disassembler(32)
                else:
                    self.assembler = self._init_x86_assembler(64)
                    self.disassembler = self._init_x86_disassembler(64)
            elif architecture == 'arm':
                if mode == 32:
                    self.assembler = self._init_arm_assembler(32)
                    self.disassembler = self._init_arm_disassembler(32)
                else:
                    self.logger.error("ARM64 mode not supported")
                    return False
            else:
                self.logger.error(f"Unsupported architecture: {architecture}")
                return False
                
            return True
            
        except Exception as e:
            self.logger.error(f"Error initializing memory manager: {e}")
            return False
            
    def patch_memory(self, address: int, data: bytes) -> bool:
        """Patch memory at specified address with data.
        
        Args:
            address: Target memory address
            data: Data to write
            
        Returns:
            bool: True if patching was successful
        """
        try:
            # Store original bytes
            self.original_bytes[address] = self.read_memory(address, len(data))
            
            # Write new data
            self.write_memory(address, data)
            
            # Store patch information
            self.patches[address] = {
                'data': data,
                'size': len(data)
            }
            
            return True
            
        except Exception as e:
            self.logger.error(f"Error patching memory: {e}")
            return False
            
    def patch_instruction(self, address: int, instruction: str) -> bool:
        """Patch memory with an assembly instruction.
        
        Args:
            address: Target memory address
            instruction: Assembly instruction to write
            
        Returns:
            bool: True if patching was successful
        """
        try:
            # Assemble instruction
            data = self.assembler.assemble(instruction)
            if not data:
                return False
                
            # Apply patch
            return self.patch_memory(address, data)
            
        except Exception as e:
            self.logger.error(f"Error patching instruction: {e}")
            return False
            
    def patch_jump(self, address: int, target: int) -> bool:
        """Create a jump instruction to target address.
        
        Args:
            address: Source address
            target: Target address
            
        Returns:
            bool: True if patching was successful
        """
        try:
            # Calculate relative offset
            offset = target - (address + 5)  # 5 bytes for JMP instruction
            
            # Create jump instruction
            instruction = f"JMP {offset}"
            return self.patch_instruction(address, instruction)
            
        except Exception as e:
            self.logger.error(f"Error patching jump: {e}")
            return False
            
    def patch_nop(self, address: int, size: int) -> bool:
        """Patch memory with NOP instructions.
        
        Args:
            address: Target memory address
            size: Number of bytes to patch
            
        Returns:
            bool: True if patching was successful
        """
        try:
            # Create NOP instruction
            nop = b'\x90'  # x86 NOP instruction
            data = nop * size
            return self.patch_memory(address, data)
            
        except Exception as e:
            self.logger.error(f"Error patching NOP: {e}")
            return False
            
    def restore_memory(self, address: int) -> bool:
        """Restore original memory content at address.
        
        Args:
            address: Target memory address
            
        Returns:
            bool: True if restoration was successful
        """
        try:
            if address in self.original_bytes:
                data = self.original_bytes[address]
                self.write_memory(address, data)
                del self.original_bytes[address]
                del self.patches[address]
                return True
            return False
            
        except Exception as e:
            self.logger.error(f"Error restoring memory: {e}")
            return False
            
    def disassemble(self, address: int, size: int) -> List[str]:
        """Disassemble memory region.
        
        Args:
            address: Start address
            size: Number of bytes to disassemble
            
        Returns:
            List[str]: List of disassembled instructions
        """
        try:
            # Read memory
            data = self.read_memory(address, size)
            
            # Disassemble
            return self.disassembler.disassemble(data)
            
        except Exception as e:
            self.logger.error(f"Error disassembling memory: {e}")
            return []
            
    def get_patches(self) -> Dict[int, Dict]:
        """Get all applied patches.
        
        Returns:
            Dict[int, Dict]: Dictionary of patches
        """
        return self.patches
        
    def get_original_bytes(self) -> Dict[int, bytes]:
        """Get all original bytes.
        
        Returns:
            Dict[int, bytes]: Dictionary of original bytes
        """
        return self.original_bytes
        
    def clear_patches(self) -> None:
        """Clear all patches and restore original memory."""
        for address in list(self.patches.keys()):
            self.restore_memory(address)
            
    def read_memory(self, address: int, size: int) -> bytes:
        """Read memory at specified address.
        
        Args:
            address: Target memory address
            size: Number of bytes to read
            
        Returns:
            bytes: Read memory content
        """
        # Implementation will be added based on platform
        pass
        
    def write_memory(self, address: int, data: bytes) -> bool:
        """Write data to memory at specified address.
        
        Args:
            address: Target memory address
            data: Data to write
            
        Returns:
            bool: True if writing was successful
        """
        # Implementation will be added based on platform
        pass
        
    def _init_x86_assembler(self, mode: int):
        """Initialize x86 assembler.
        
        Args:
            mode: Architecture mode (32 or 64)
            
        Returns:
            Assembler instance
        """
        # Implementation will be added based on assembler library
        pass
        
    def _init_x86_disassembler(self, mode: int):
        """Initialize x86 disassembler.
        
        Args:
            mode: Architecture mode (32 or 64)
            
        Returns:
            Disassembler instance
        """
        # Implementation will be added based on disassembler library
        pass
        
    def _init_arm_assembler(self, mode: int):
        """Initialize ARM assembler.
        
        Args:
            mode: Architecture mode (32 or 64)
            
        Returns:
            Assembler instance
        """
        # Implementation will be added based on assembler library
        pass
        
    def _init_arm_disassembler(self, mode: int):
        """Initialize ARM disassembler.
        
        Args:
            mode: Architecture mode (32 or 64)
            
        Returns:
            Disassembler instance
        """
        # Implementation will be added based on disassembler library
        pass 